refactor: 안쓰는거 좀 지우고 이름 바꿈
