+++ frontend/src/services/getUserInfo.ts
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 import { API } from 'common/constants';
 import { UserInfoType } from 'types/user';
-import { throwAxiosFtError } from '../utils/error/throwAxiosFtError';
+import { throwApiError } from '../utils/error';
 
 interface Props {
   userId?: string;
@@ -14,5 +14,5 @@ export function getUserInfo({ userId }: Props): Promise<UserInfoType> {
       withCredentials: true,
     })
     .then(({ data }) => data)
-    .catch(throwAxiosFtError);
+    .catch(throwApiError);
 }
