+++ frontend/src/services/postRegister.ts
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 import { API } from 'common/constants';
 import { UserInfoType } from '../types/user';
-import { throwAxiosFtError } from '../utils/error/throwAxiosFtError';
+import { throwApiError } from '../utils/error';
 
 // TODO: refactor me!! - by ycha
 export async function postRegister(nickname: string, avatar: string): Promise<UserInfoType> {
@@ -18,5 +18,5 @@ export async function postRegister(nickname: string, avatar: string): Promise<Us
       },
     )
     .then(({ data }) => data)
-    .catch(throwAxiosFtError);
+    .catch(throwApiError);
 }
