+++ frontend/src/utils/error/ApiError.ts
@@ -1,0 +1,9 @@
+export class ApiError extends Error {
+  statusCode: number;
+
+  constructor(statusCode: number, message: string) {
+    super(message);
+    this.name = 'ApiError';
+    this.statusCode = statusCode;
+  }
+}
