--- frontend/src/utils/error/ftError.ts
@@ -1,11 +1,0 @@
-import { FtErrorResponse } from './ftErrorResponse';
-
-export class FtError extends Error {
-  statusCode: number;
-
-  constructor(errorResponse: FtErrorResponse) {
-    super(errorResponse.message);
-    this.name = `FtError(${errorResponse.error})`;
-    this.statusCode = errorResponse.statusCode;
-  }
-}
