--- frontend/src/utils/error/ftErrorResponse.ts
@@ -1,5 +1,0 @@
-export interface FtErrorResponse {
-  error: string;
-  message: string;
-  statusCode: number;
-}
