--- frontend/src/utils/error/isFtError.ts
@@ -1,9 +1,0 @@
-import { FtError } from './ftError';
-
-export function isFtError(error: unknown): error is FtError {
-  if (error instanceof FtError) {
-    return true;
-  }
-
-  return false;
-}
