--- frontend/src/utils/error/throwAxiosFtError.ts
@@ -1,10 +1,0 @@
-import { FtError } from './ftError';
-import { isAxiosFtErrorResponse } from './isAxiosFtErrorResponse';
-
-export function throwAxiosFtError(error: unknown): never {
-  if (isAxiosFtErrorResponse(error)) {
-    throw new FtError(error.response.data);
-  }
-
-  throw error;
-}
